+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - quarkus-version set to 3.31.2 across all the modules <https://github.com/fugerit-org/fj-doc/issues/618>
 
+### Fixed
+
+- junit for JSON and YAML conversion and JDK 25 <https://github.com/fugerit-org/fj-doc/issues/620>
+
 ## [8.18.1] - 2026-02-04
 
 ### Fixed